+++ jtransc-rt/src/java/lang/Object.java
@@ -65,6 +65,7 @@ import java.lang.reflect.Field;
 @JTranscAddFile(target = "as3", priority = -1, process = true, src = "as3/Main.xml", dst = "Main.xml")
 @JTranscAddLibraries(target = "cpp", value = "gc")
 @JTranscAddIncludes(target = "cpp", value = "gc_cpp.h")
+@JTranscAddDefines(target = "cpp", value = "GC_NOT_DLL=1")
 @JTranscAddMembers(target = "d", value = {
 	"core.sync.mutex.Mutex __d_mutex = null;",
 })